added modular loading for interactions
