+++ src/bot.ts
@@ -1,4 +1,10 @@
-import { Client, Collection, IntentsBitField, Snowflake } from "discord.js";
+import {
+  BaseInteraction,
+  Client,
+  Collection,
+  IntentsBitField,
+  Snowflake,
+} from "discord.js";
 import { DB } from "./db.js";
 import { Command } from "./modules/command.js";
 import { FakeRandom, Random } from "./modules/random.js";
@@ -19,6 +25,10 @@ export class Bot extends Client {
   }
   prefix: string = settings.PREFIX;
   commands: Collection<string, Command> = new Collection();
+  interactions: Collection<
+    string,
+    (client: Bot, interaction: BaseInteraction, args: string[]) => void
+  > = new Collection();
   commandusage: Map<Snowflake, Array<number>> = new Map();
   db: DB = new DB();
   languages: Map<string, ILanguage> = new Map();
